Refactors URL extraction script to ESM
Migrates the URL extraction script to ESM for improved module handling.

Updates the Docker configuration to execute the new .mjs script.
